Update backend 24.6.2022

+++ CRMBug-BE/ApplicationCore/Interfaces/DL/IDLProject.cs
@@ -9,6 +9,6 @@ namespace ApplicationCore.Interfaces.DL
 {
   public interface IDLProject : IDLBase<Project>
   {
-    bool InviteUser(int projectID, List<int> userIDs);
+    bool InviteUser(long projectID, List<long> userIDs);
   }
 }
